requirements.txt
fastapi==0.75.0
uvicorn==0.17.6
web3==5.29.1
ccxt==1.83.56
python-dotenv==0.19.2